requirements.txt
librosa==0.8.1
pandas==1.2.5
pytorch==1.9.0
pytorch-cluster==1.5.9
pytorch-geometric==1.7.2
pytorch-scatter==2.0.7
pytorch-sparse==0.6.10
pytorch-spline-conv==1.2.1
scikit-learn
scipy
torchaudio=0.9.0
tqdm
yaml
